+++ api/main.py
@@ -1,48 +1,47 @@
-import os
-from pprint import pprint
+import logging
 
+from peewee import DoesNotExist
 from fastapi import FastAPI, HTTPException
 from typing import Optional
 
-from database.tumread import TUMReadDatabase
+from database.database import init_db, Module, Degree, Mapping
 
 app = FastAPI()
 
-database = TUMReadDatabase()
+init_db()
 
 @app.get("/")
 def get_counts():
     return {
-        "number_of_degrees": database.get_number_of_degrees(),
-        "number_of_modules": database.get_number_of_modules(),
-        "number_of_mappings": database.get_number_of_mappings()
+        "number_of_degrees": Degree.select().count(),
+        "number_of_modules": Module.select().count(),
+        "number_of_mappings": Mapping.select().count()
     }
 
 
 @app.get("/degree")
 def get_degree(degree_id: int):
-    degree_info = database.get_degree_info(degree_id)
-    if not degree_info:
+    try:
+        return Degree.get_by_id(degree_id)
+    except DoesNotExist:
         raise HTTPException(status_code=404, detail="Degree not found")
-    return degree_info
-
 
 @app.get("/degrees")
 def get_degrees():
-    return list(database.get_all_degrees())
+    return list(Degree.select().count())
 
 
 @app.get("/module")
 def get_module(module_id: int):
-    module_info = database.get_module_info(module_id)
-    if not module_info:
+    try:
+        return Module.get_by_id(module_id)
+    except DoesNotExist:
         raise HTTPException(status_code=404, detail="Module not found")
-    return module_info
 
 @app.get("/modules_of_degree")
 def get_modules_of_degree(degree_id: int, valid_from: Optional[str] = None, valid_to: Optional[str] = None,
                           degree_version: Optional[str] = None):
-    modules = database.get_modules_of_degree(degree_id, valid_from, valid_to, degree_version)
+    modules = get_modules_of_degree(degree_id, valid_from, valid_to, degree_version)
     if not modules:
         raise HTTPException(status_code=404, detail="No modules found for the given degree")
     return list(modules)
